DISPATCH-89: remove the check for exchange in the forwarder
Rather than check for the presence of an exchange each time a message
is forwarded, have the exchange override the qdr_address_t's forwarder
pointer.

When I say "zero performance impact to existing configurations", I
mean "zero performance impact to existing configurations".

Word.
